Rename menu to main (typo).
