+++ src/components/Main/Carousel/Carousel.jsx
