+++ src/components/Main/Carousel/CarouselControl.jsx
