+++ src/components/Main/Icons/TravelTypesIcons.jsx
