+++ src/tests/components/Menu/IconsMapping.test.js
